feat: Support Author aliases
People don't always use the same git `user.name` for their commit
messages.

This feature adds an `aliases` config settings a desired author name can
be associated with multiple commit author names. Whenever the commit
author name is discovered in the Gource log, it will be converted to the
alias before being rendered in the visualization.